+++ src/test/java/com/guillermomolina/protos/execution/ProtosCoreBootstrapTest.java
@@ -132,6 +132,7 @@ class ProtosCoreBootstrapTest {
                         "LineTooLong",
                         "Array",
                         "String",
+                        "Encoding",
                         "Map",
                         "IdentityMap",
                         "Path",